requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
